+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [#7528](https://github.com/rubocop-hq/rubocop/pull/7528): Add new `Lint/NonDeterministicRequireOrder` cop. ([@mangara][])
 * [#7559](https://github.com/rubocop-hq/rubocop/pull/7559): Add `EnforcedStyleForExponentOperator` parameter to `Layout/SpaceAroundOperators` cop. ([@khiav223577][])
+* [#7565](https://github.com/rubocop-hq/rubocop/pull/7565): Add support for `begin` to `Layout/EndAlignment` cop. ([@jaredbeck][])
 
 ### Bug fixes